number format pada halaman edit sparepart dan perbaikan pada halaman tambah customer dan kendaraan baru

+++ database/migrations/2024_11_05_215541_create_spareparts_table.php
@@ -13,9 +13,12 @@ class CreateSparepartsTable extends Migration
             $table->string('nama_sparepart');
             $table->string('spek');
             $table->integer('jumlah');
-            $table->decimal('harga_beli', 15, 2);
-            $table->decimal('harga_jual', 15, 2);
-            $table->decimal('keuntungan', 15, 2)->computedAs('harga_jual - harga_beli');
+            // $table->decimal('harga_beli', 15, 2);
+            // $table->decimal('harga_jual', 15, 2);
+            // $table->decimal('keuntungan', 15, 2)->computedAs('harga_jual - harga_beli');
+            $table->integer('harga_beli');
+            $table->integer('harga_jual');
+            $table->integer('keuntungan')->computedAs('harga_jual - harga_beli');
             $table->date('tanggal_masuk');
             $table->date('tanggal_keluar')->nullable();
             $table->text('deskripsi')->nullable();
